requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
sentence-transformers==2.2.2
pypdf==3.17.1
scikit-learn==1.3.2
numpy==1.24.3
torch==2.1.0
transformers==4.35.0
spacy==3.7.2
faiss-cpu==1.7.4
pandas==2.0.3

